pyproject.toml
[tool.poetry]
name = "custom-kfp-components"
version = "0.1.0"
description = ""
authors = ["Mónica Milán <monicamilan@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python =  ">=3.11,<3.13.0"
kfp = "^2.7.0"
google-cloud-aiplatform = "^1.47.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pandas = "^2.2.2"
scikit-learn = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
